refactor(vscode): split extension.ts into focused modules
Split the 1437-line VS Code extension entry point into 4 modules:
- extension.ts (147 lines): activate/deactivate + command registration
- view-provider.ts (346 lines): FrontAgentViewProvider class
- settings-resolve.ts (126 lines): configuration resolution and setup wizard
- webview-html.ts (834 lines): webview HTML/CSS/JS template

Closes #31
